\name{ex1.data}
\alias{ex1.data}

\docType{data}
\title{
	Example data frame
}
\description{
   Simulated data from a Gaussian contamination model
}
\usage{data(ex1.data)}
\format{
  A data frame with 500 observations and 2 variables (\code{X1},\code{Y1}).
  \describe{
    \item{\code{X1}}{error-free variable (numeric)}
    \item{\code{Y1}}{contaminated variable (numeric)}
  }
}
\details{
Data have been generated by a Gaussian  model. The
 variable Y1  has been contaminated with parameters B=(-0.26, 1.26), sigma=1.21, w=0.05, lambda=10.
}

\examples{
data(ex1.data)

}
\keyword{datasets}
